python_regression_projects/food_delivery_time_prediction/server/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
pymongo
python-dotenv
gunicorn